hotfix - controller input for alphabets

+++ DSA/LAB 2/Collinearity.c
@@ -12,8 +12,8 @@ int collinearity()
     printf("Enter co-ordinates of thirs point (x3,y3): ");
     scanf("%d%d", &x3, &y3);
 
-    int m1 = y2 - y1 / x2 - x1;
-    int m2 = y3 - y2 / x3 - x2;
+    float m1 = y2 - y1 / x2 - x1;
+    float m2 = y3 - y2 / x3 - x2;
 
     if (m1 == m2)
         printf("\nPoints are collinear");
